Seperated UI and Game Conditions into Controllers with their own file
