+++ src/gamecontroller.cpp
@@ -1,0 +1,28 @@
+#include "gamecontroller.h"
+
+GameController::GameController() {}
+
+void GameController::StartGame()
+{
+
+}
+
+void GameController::ResetGame()
+{
+
+}
+
+void GameController::PlayerHit()
+{
+
+}
+
+void GameController::PlayerStand()
+{
+
+}
+
+void GameController::DealerTurn()
+{
+
+}
