+++ src/uicontroller.cpp
@@ -1,0 +1,6 @@
+#include "uicontroller.h"
+
+uicontroller::uicontroller()
+{
+
+}
